test(storage): run integration suite against real Postgres in CI
Parameterise the mcs-storage integration suite so the same test bodies run
on both backends. A new `tests::harness` helper, `connect_test_storage`,
reads `MCS_TEST_DATABASE_URL` (defaulting to `sqlite::memory:`) and returns
an isolated, migrated `SqlxStorage`:

- SQLite: a private in-memory database per test (torn down with the pool).
- Postgres: a private, uniquely-named schema per test, with every pool
  connection pinning `search_path` to it via `after_connect`. The schema is
  dropped on teardown. Per-test uniqueness keeps a shared server isolated
  with no cross-test interference.

The old SQLite-only `sqlx_sqlite.rs` becomes the backend-agnostic
`sqlx_backend.rs`, and its `storage()` helper now goes through the harness.

Fix portability bugs the Postgres run surfaced: integer columns the code
binds/reads as `i64` (`ply`, clocks, `rated`) move from `INTEGER` to
`BIGINT`, and Glicko-2 rating columns read as `f64` move from `REAL` to
`DOUBLE PRECISION`. Postgres `INTEGER`/`REAL` are 4-byte and would not
decode as `i64`/`f64`; both widened types keep SQLite's affinity behaviour
unchanged.

Add a `postgres integration tests` CI job with a `postgres:16` service that
sets `MCS_TEST_DATABASE_URL` and runs the suite under
`--no-default-features --features postgres --test-threads=1`. The existing
`ci` job keeps running the suite on SQLite.

+++ crates/mcs-storage/migrations/0002_ratings.sql
@@ -2,8 +2,10 @@
 --
 -- The primary key is the composite (user_id, variant_id) pair; there is
 -- exactly one rating row per player per game variant.  All three Glicko-2
--- parameters are stored as REAL so they survive a round-trip through both
--- SQLite (which uses IEEE 754 doubles) and PostgreSQL (DOUBLE PRECISION).
+-- parameters are stored as DOUBLE PRECISION so they survive a round-trip
+-- through both SQLite (IEEE 754 doubles via REAL affinity) and PostgreSQL
+-- (an 8-byte float decoded as `f64`). Postgres `REAL` is only 4 bytes and
+-- would not decode as `f64`, so the wider type is required for portability.
 --
 -- No foreign-key constraint on user_id is added: variants are application-
 -- level strings, and the users table may not always be in the same schema
@@ -12,9 +14,9 @@
 CREATE TABLE IF NOT EXISTS ratings (
     user_id    TEXT    NOT NULL,
     variant_id TEXT    NOT NULL,
-    value      REAL    NOT NULL,
-    deviation  REAL    NOT NULL,
-    volatility REAL    NOT NULL,
+    value      DOUBLE PRECISION NOT NULL,
+    deviation  DOUBLE PRECISION NOT NULL,
+    volatility DOUBLE PRECISION NOT NULL,
     PRIMARY KEY (user_id, variant_id)
 );
