+++ crates/mcs-storage/migrations/0006_challenges.sql
@@ -7,8 +7,9 @@
 -- Portability note: as with the earlier migrations, this DDL runs unchanged on
 -- both SQLite and PostgreSQL, so it sticks to the lowest common denominator —
 -- TEXT for ids, enum discriminants, variant ids, and JSON-encoded value objects
--- (the time control); an INTEGER boolean (0 = casual, 1 = rated) for `rated`,
--- matching how the codebase stores other small integers; and RFC 3339 TEXT for
+-- (the time control); a BIGINT boolean (0 = casual, 1 = rated) for `rated`,
+-- matching how the codebase stores other small integers (bound and read as
+-- 8-byte `i64`, which maps onto Postgres `BIGINT`); and RFC 3339 TEXT for
 -- timestamps. The `game_id` is NULL until the challenge is accepted.
 
 CREATE TABLE IF NOT EXISTS challenges (
@@ -18,7 +19,7 @@ CREATE TABLE IF NOT EXISTS challenges (
     variant_id       TEXT NOT NULL,
     -- JSON-encoded `mcs_domain::TimeControl`.
     time_control     TEXT NOT NULL,
-    rated            INTEGER NOT NULL DEFAULT 1,
+    rated            BIGINT NOT NULL DEFAULT 1,
     color_preference TEXT NOT NULL,
     -- One of: pending, accepted, declined, canceled.
     status           TEXT NOT NULL,
